+++ graph/graph-adapter-input-rest-spring-mvc/src/test/kotlin/eu/tib/orkg/prototype/contenttypes/VisualizationControllerUnitTest.kt
@@ -131,7 +131,9 @@ internal class VisualizationControllerUnitTest : RestDocsTest("visualizations") 
                     requestParameters(
                         parameterWithName("title").description("Optional filter for the title of the visualization. Uses exact matching.").optional(),
                         parameterWithName("visibility").description("""Optional filter for visibility. Either of "listed", "featured", "unlisted" or "deleted".""").optional(),
-                        parameterWithName("created_by").description("Optional filter for research field id.").optional(),
+                        parameterWithName("created_by").description("Optional filter for the UUID of the user or service who created the visualization.").optional(),
+                        parameterWithName("research_field").description("Optional filter for research field id.").optional(),
+                        parameterWithName("include_subfields").description("Optional flag for whether subfields are included in the search or not.").optional(),
                     )
                 )
             )
@@ -203,5 +205,37 @@ internal class VisualizationControllerUnitTest : RestDocsTest("visualizations") 
             .andExpect(jsonPath("$.status").value(HttpStatus.BAD_REQUEST.value()))
             .andExpect(jsonPath("$.path").value("/api/visualizations"))
             .andExpect(jsonPath("$.message").value(exception.message))
+    }
+
+    @Test
+    fun `Given several visualizations, when they are fetched by visibility and research field id, then status is 200 OK and visualizations are returned`() {
+        val visualizations = listOf(createDummyVisualization())
+        val researchFieldId = ThingId("Science")
+        every {
+            visualizationService.findAllByResearchFieldAndVisibility(
+                researchFieldId = researchFieldId,
+                visibility = VisibilityFilter.ALL_LISTED,
+                includeSubfields = true,
+                pageable = any()
+            )
+        } returns PageImpl(visualizations, PageRequest.of(0, 5), 1)
+
+        get("/api/visualizations?research_field=$researchFieldId&visibility=ALL_LISTED&include_subfields=true")
+            .accept(VISUALIZATION_JSON_V2)
+            .contentType(VISUALIZATION_JSON_V2)
+            .perform()
+            .andExpect(status().isOk)
+            .andExpect(jsonPath("$.content", hasSize<Int>(1)))
+            .andExpect(jsonPath("$.number").value(0)) // page number
+            .andExpect(jsonPath("$.totalElements").value(1))
+
+        verify(exactly = 1) {
+            visualizationService.findAllByResearchFieldAndVisibility(
+                researchFieldId = researchFieldId,
+                visibility = VisibilityFilter.ALL_LISTED,
+                includeSubfields = true,
+                pageable = any()
+            )
+        }
     }
 }