+++ graph/graph-application/src/main/kotlin/eu/tib/orkg/prototype/contenttypes/api/visualization-queries.kt
@@ -1,7 +1,7 @@
 package eu.tib.orkg.prototype.contenttypes.api
 
-import eu.tib.orkg.prototype.contenttypes.domain.model.Visualization
 import eu.tib.orkg.prototype.community.domain.model.ContributorId
+import eu.tib.orkg.prototype.contenttypes.domain.model.Visualization
 import eu.tib.orkg.prototype.statements.api.VisibilityFilter
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
 import java.util.*
@@ -14,4 +14,10 @@ interface RetrieveVisualizationUseCase {
     fun findAllByTitle(title: String, pageable: Pageable): Page<Visualization>
     fun findAllByContributor(contributorId: ContributorId, pageable: Pageable): Page<Visualization>
     fun findAllByVisibility(visibility: VisibilityFilter, pageable: Pageable): Page<Visualization>
+    fun findAllByResearchFieldAndVisibility(
+        researchFieldId: ThingId,
+        visibility: VisibilityFilter,
+        includeSubfields: Boolean,
+        pageable: Pageable
+    ): Page<Visualization>
 }