+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (See: [!668](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/668))
 - It is now possible to filter comparisons by `visibility`, `research_field` and optionally `include_subfields`
   (See: [!669](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/669))
+- It is now possible to filter visualizations by `visibility`, `research_field` and optionally `include_subfields`
+  (See: [!670](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/670))
 
 ## [0.39.0] - 2023-10-16
 ### Added